dom/tests/mochitest/bugs/test_bug581072.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=581072
-->
<head>
  <title>Test for Bug 581072</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=581072">Mozilla Bug 581072</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();
  
/** Test for Bug 581072 **/
var longTimerFired = 0;

// Set up a one-hour timeout
setTimeout(function() { longTimerFired = true; }, 3600000);

// Trigger suspend and resume timeouts
(function() {
  var utils = SpecialPowers.getDOMWindowUtils(window);
  utils.suspendTimeouts();  
  utils.resumeTimeouts();  
})()

// Now set up another timeout which should fire before the one-hour one
setTimeout(function() {
  is(longTimerFired, false, "One-hour timer should not fire before our 0ms one");
  SimpleTest.finish();
}, 0);



</script>
</pre>
</body>
</html>
